Added image + package reorganisation + pytest now tests mpl correctly
