+++ pyproject.toml
@@ -39,6 +39,7 @@ matplotlib = "~3.3.2"
 moderngl = "^5.6.2"
 multiprocess = "^0.70.11"
 numpy = "~1.19.5"
+Pillow = "^8.1.0"
 PySide2 = "^5.15.2"
 Shapely = {extras = ["vectorized"], version = "~1.7.1"}
 scipy = "~1.5.4"
